LIVE-1575: add Headline story
